backend/pyproject.toml
[project]
name = "ltx-desktop-backend"
version = "1.0.0"
description = "LTX-2 Video Generation Backend"
requires-python = ">=3.12"
dependencies = [
    "pillow>=10.3.0",
    "torch>=2.3.0",
    "huggingface-hub>=0.23.0",
    "tqdm>=4.66.0",
    "pynvml>=11.5.0; sys_platform != 'darwin'",
    "pydantic>=2.7.0",
    # LTX-2 inference dependencies
    "ltx-core",
    "ltx-pipelines",
    "diffusers>=0.36.0",
    "ftfy>=6.0.0",
    "imageio>=2.37.2",
    "imageio-ffmpeg>=0.6.0",
    "peft>=0.13.2",
    "protobuf>=3.20.0",
    "transformers>=4.52,<5",
    "sentencepiece>=0.1.99",
    "sageattention>=1.0.0; sys_platform != 'darwin'",
    "opencv-python-headless>=4.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "websocket-client>=1.6.0",
    "triton-windows; sys_platform == 'win32'",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]
torchaudio = { index = "pytorch-cu128" }
diffusers = { git = "https://github.com/huggingface/diffusers.git", rev = "01de02e8b4f2cc91df4f3e91cb6535ebcbeb490c" }
ltx-core = { git = "https://github.com/Lightricks/LTX-2.git", subdirectory = "packages/ltx-core", rev = "00dc53d3f81c405932f9f16d9c57557de411e702" }
ltx-pipelines = { git = "https://github.com/Lightricks/LTX-2.git", subdirectory = "packages/ltx-pipelines", rev = "00dc53d3f81c405932f9f16d9c57557de411e702" }

[project.optional-dependencies]
test = ["pytest>=8.0", "requests>=2.31", "httpx>=0.27"]
dev = ["pyright>=1.1.380", "debugpy>=1.8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
